# internal/web/apidoc/docs/swagger.yaml
basePath: /
definitions:
  apidoc.JsonResponseError:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  apidoc.JsonResponseOk:
    properties:
      data: {}
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.AcceptAccessRequest:
    properties:
      budgetId:
        type: string
    type: object
  model.AcceptAccessResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.MetaResult'
        type: array
    type: object
...
  model.AccessResult:
    properties:
      isAccepted:
        type: integer
      role:
        type: string
      user:
        $ref: '#/definitions/model.UserResult'
    type: object
...
  model.BudgetFolderResult:
    properties:
      id:
        type: string
      name:
        type: string
      position:
        type: integer
    type: object
...
  model.ChangeElementCurrencyRequest:
    properties:
      budgetId:
        type: string
      currencyId:
        type: string
      elementId:
        type: string
    type: object
  model.ChangeElementCurrencyResult:
    type: object
  model.ChildElementResult:
    properties:
      budgetSpent:
        type: string
      icon:
        type: string
      id:
        type: string
      isArchived:
        type: integer
      name:
        type: string
      ownerUserId:
        type: string
      spent:
        type: string
      type:
        type: integer
    type: object
...
  model.CreateBudgetFolderRequest:
    properties:
      budgetId:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  model.CreateBudgetFolderResult:
    properties:
      item:
        $ref: '#/definitions/model.BudgetFolderResult'
    type: object
...
  model.CreateEnvelopeRequest:
    properties:
      budgetId:
        type: string
      categories:
        items:
          type: string
        type: array
      currencyId:
        type: string
      folderId:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  model.CreateEnvelopeResult:
    properties:
      item:
        $ref: '#/definitions/model.ParentElementResult'
    type: object
...
  model.DeclineAccessRequest:
    properties:
      budgetId:
        type: string
    type: object
  model.DeclineAccessResult:
    type: object
...
  model.DeleteEnvelopeRequest:
    properties:
      budgetId:
        type: string
      id:
        type: string
    type: object
  model.DeleteEnvelopeResult:
    type: object
  model.DeleteFolderRequest:
    properties:
      budgetId:
        type: string
      id:
        type: string
    type: object
  model.DeleteFolderResult:
    type: object
...
  model.ExcludeAccountRequest:
    properties:
      accountId:
        type: string
      id:
        type: string
    type: object
  model.ExcludeAccountResult:
    properties:
      item:
        $ref: '#/definitions/model.MetaResult'
    type: object
...
  model.GrantAccessRequest:
    properties:
      budgetId:
        type: string
      role:
        type: string
      userId:
        type: string
    type: object
  model.GrantAccessResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.MetaResult'
        type: array
    type: object
...
  model.IncludeAccountRequest:
    properties:
      accountId:
        type: string
      id:
        type: string
    type: object
  model.IncludeAccountResult:
    properties:
      item:
        $ref: '#/definitions/model.MetaResult'
    type: object
...
  model.MetaResult:
    properties:
      access:
        items:
          $ref: '#/definitions/model.AccessResult'
        type: array
      currencyId:
        type: string
      id:
        type: string
      name:
        type: string
      ownerUserId:
        type: string
      startedAt:
        type: string
    type: object
  model.MoveElementListItem:
    properties:
      folderId:
        type: string
      id:
        type: string
      position:
        type: integer
    type: object
  model.MoveElementListRequest:
    properties:
      budgetId:
        type: string
      items:
        items:
          $ref: '#/definitions/model.MoveElementListItem'
        type: array
    type: object
  model.MoveElementListResult:
    type: object
...
  model.OrderBudgetFolderListRequest:
    properties:
      budgetId:
        type: string
      items:
        items:
          $ref: '#/definitions/model.OrderFolderListItem'
        type: array
    type: object
  model.OrderBudgetFolderListResult:
    type: object
...
  model.OrderFolderListItem:
    properties:
      id:
        type: string
      position:
        type: integer
    type: object
...
  model.ParentElementResult:
    properties:
      available:
        type: string
      budgetSpent:
        type: string
      budgeted:
        type: string
      children:
        items:
          $ref: '#/definitions/model.ChildElementResult'
        type: array
      currencyId:
        type: string
      folderId:
        type: string
      icon:
        type: string
      id:
        type: string
      isArchived:
        type: integer
      name:
        type: string
      ownerUserId:
        type: string
      position:
        type: integer
      spent:
        type: string
      type:
        type: integer
    type: object
...
  model.RevokeAccessRequest:
    properties:
      budgetId:
        type: string
      userId:
        type: string
    type: object
  model.RevokeAccessResult:
    type: object
...
  model.SetLimitRequest:
    properties:
      amount:
        type: string
      budgetId:
        type: string
      elementId:
        type: string
      period:
        type: string
    type: object
  model.SetLimitResult:
    type: object
...
  model.UpdateBudgetFolderRequest:
    properties:
      budgetId:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  model.UpdateBudgetFolderResult:
    properties:
      item:
        $ref: '#/definitions/model.BudgetFolderResult'
    type: object
...
  model.UpdateEnvelopeRequest:
    properties:
      budgetId:
        type: string
      categories:
        items:
          type: string
        type: array
      currencyId:
        type: string
      icon:
        type: string
      id:
        type: string
      isArchived:
        type: integer
      name:
        type: string
    type: object
  model.UpdateEnvelopeResult:
    properties:
      item:
        $ref: '#/definitions/model.ParentElementResult'
    type: object
...
  model.UserResult:
    properties:
      avatar:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Self-hosted personal finance and budgeting API
  title: Econumo API
  version: 1.0.0
paths:
...
  /api/v1/budget/accept-access:
    post:
      consumes:
      - application/json
      parameters:
      - description: Accept access
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AcceptAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.AcceptAccessResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Accept budget access
      tags:
      - Budget
  /api/v1/budget/change-element-currency:
    post:
      consumes:
      - application/json
      parameters:
      - description: Change element currency
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ChangeElementCurrencyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.ChangeElementCurrencyResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Change a budget element's currency
      tags:
      - Budget
...
  /api/v1/budget/create-envelope:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create envelope
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateEnvelopeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.CreateEnvelopeResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Create an envelope
      tags:
      - Budget
  /api/v1/budget/create-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateBudgetFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.CreateBudgetFolderResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Create a budget folder
      tags:
      - Budget
  /api/v1/budget/decline-access:
    post:
      consumes:
      - application/json
      parameters:
      - description: Decline access
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeclineAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.DeclineAccessResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Decline budget access
      tags:
      - Budget
...
  /api/v1/budget/delete-envelope:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete envelope
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeleteEnvelopeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.DeleteEnvelopeResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Delete an envelope
      tags:
      - Budget
  /api/v1/budget/delete-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Delete folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeleteFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.DeleteFolderResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Delete a budget folder
      tags:
      - Budget
  /api/v1/budget/exclude-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Exclude account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ExcludeAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.ExcludeAccountResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Exclude an account from a budget
      tags:
      - Budget
...
  /api/v1/budget/grant-access:
    post:
      consumes:
      - application/json
      parameters:
      - description: Grant access
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.GrantAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.GrantAccessResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Grant budget access
      tags:
      - Budget
  /api/v1/budget/include-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Include account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.IncludeAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.IncludeAccountResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Include an account in a budget
      tags:
      - Budget
  /api/v1/budget/move-element-list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Move elements
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MoveElementListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.MoveElementListResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Move/reorder budget elements
      tags:
      - Budget
  /api/v1/budget/order-folder-list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order folders
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.OrderBudgetFolderListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderBudgetFolderListResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Reorder budget folders
      tags:
      - Budget
...
  /api/v1/budget/revoke-access:
    post:
      consumes:
      - application/json
      parameters:
      - description: Revoke access
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RevokeAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.RevokeAccessResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Revoke budget access
      tags:
      - Budget
  /api/v1/budget/set-limit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Set limit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SetLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.SetLimitResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Set a budget element limit
      tags:
      - Budget
...
  /api/v1/budget/update-envelope:
    post:
      consumes:
      - application/json
      parameters:
      - description: Update envelope
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateEnvelopeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateEnvelopeResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Update an envelope
      tags:
      - Budget
  /api/v1/budget/update-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Update folder
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateBudgetFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateBudgetFolderResult'
              type: object
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
      security:
      - Bearer: []
      summary: Update a budget folder
      tags:
      - Budget
...
securityDefinitions:
  Bearer:
    description: 'Opaque access token. Format: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"